requirements.txt
# Core Processing
kafka-python==2.0.2
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Machine Learning
scikit-learn==1.3.0

# Database Clients
influxdb-client==1.37.0
psycopg2-binary==2.9.7

# Audio Processing (opcional para análisis avanzado)
librosa==0.10.1
soundfile==0.12.1

# Dashboard & Visualization
streamlit==1.27.0
plotly==5.16.1
matplotlib==3.7.2
seaborn==0.12.2

# Configuration & Utils
pyyaml==6.0.1
python-dotenv==1.0.0
python-telegram-bot==20.4  # Para alertas por Telegram

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Development
black==23.7.0
flake8==6.1.0
mypy==1.5.1

# Monitoring (opcional)
prometheus-client==0.17.1

# Data Validation
pydantic==2.3.0